reorganised sorting classes
