+++ src/InsertionSort.java
@@ -1,0 +1,42 @@
+package com.matomatical.ads;
+
+import java.util.Comparator;
+
+public class InsertionSort extends Sort {
+
+	// this class should not be instantiated
+	private InsertionSort() {}
+
+	/**
+	 * Sorts the elements of A by a custom comparator using insertion sort
+	 * @param A generic array of objects to sort
+	 * @param comparator a comparator that can compare the objects in the array
+	 */
+	public static <T> void insertionSort(T[] A, Comparator<T> comparator) {
+		// for every item in the array
+		for(int i = 1; i < A.length; i++){
+			
+			// pass through the first i-1 items looking for the right place
+			for(int j = i - 1; j >= 0; j--){
+				
+				if (less(A[j+1], A[j], comparator)) {
+					// as long as the item is above its correct place,
+					// continue to swap the item along the array
+					swap(A, j, j+1);
+				} else {
+					// when we find something the item isnt smaller than,
+					// we can finish inserting early
+					break;
+				}
+			}
+		}
+	}
+
+	/**
+	 * Sorts the elements of A by their compareTo method, using insertion sort
+	 * @param A generic array of objects to sort
+	 */
+	public static <T extends Comparable<T>> void insertionSort(T[] A) {
+		insertionSort(A, new ComparableComparator<T>());
+	}
+}
